new update

+++ src/pages/register/Register.jsx
@@ -1,6 +1,5 @@
- 
 import React, { useState } from "react";
-import {Link} from "react-router-dom";
+import { Link } from "react-router-dom";
 import axios from "axios";
 
 const Register = () => {
@@ -10,17 +9,24 @@ const Register = () => {
     fullName: "",
     email: "",
     password: "",
+    confirmPassword: "",
   });
 
   const handelSignUp = async () => {
     try {
-      const response = await axios.post("https://drcbd-backend.onrender.com/user/signup", user);
-      console.log(response.data)
+      if (user.password !== user.confirmPassword) {
+        return alert("Password did not match");
+      }
+      const response = await axios.post(
+        "https://drcbd-backend.onrender.com/user/signup",
+        user
+      );
+      console.log(response.data);
       localStorage.setItem("token", response.data);
       setLogin(true);
-  } catch (error) {
+    } catch (error) {
       console.log("Signup failed", error.message);
-  }
+    }
   };
 
   return (
@@ -46,23 +52,23 @@ const Register = () => {
             alt="/"
             style={{ width: "100%", objectFit: "contain", height: "30em" }}
           />
-          <Link to="/" style={{ color: "#99a79f" ,width: "35%",}}>
-          <button
-            style={{
-              marginTop: "1em",
-              fontSize: "30px",
-              textAlign: "center",
-              width: "100%",
-              padding: "10px",
-              border: "none",
-              background: "#0b4640",
-              color: "#fff",
-              fontWeight: "bold",
-            }}
-            onClick={() => setLogin(true)}
-          >
-            Go Home
-          </button>
+          <Link to="/" style={{ color: "#99a79f", width: "35%" }}>
+            <button
+              style={{
+                marginTop: "1em",
+                fontSize: "30px",
+                textAlign: "center",
+                width: "100%",
+                padding: "10px",
+                border: "none",
+                background: "#0b4640",
+                color: "#fff",
+                fontWeight: "bold",
+              }}
+              onClick={() => setLogin(true)}
+            >
+              Go Home
+            </button>
           </Link>
         </>
       ) : (
@@ -89,6 +95,7 @@ const Register = () => {
               padding: "5px",
               fontSize: 20,
               borderRadius: 10,
+              border:"1px solid grey"
             }}
             onChange={(e) => {
               setUser({ ...user, fullName: e.target.value });
@@ -110,6 +117,7 @@ const Register = () => {
               padding: "5px",
               fontSize: 20,
               borderRadius: 10,
+              border:"1px solid grey"
             }}
             onChange={(e) => {
               setUser({ ...user, email: e.target.value });
@@ -132,12 +140,13 @@ const Register = () => {
               fontSize: 16,
               borderRadius: 10,
               marginBottom: "1em",
+              border:"1px solid grey"
             }}
             onChange={(e) => {
               setUser({ ...user, password: e.target.value });
             }}
           />
-           <p
+          <p
             style={{
               width: "35%",
               fontSize: "16px",
@@ -145,7 +154,7 @@ const Register = () => {
               fontWeight: "bold",
             }}
           >
-           Confirm Password
+            Confirm Password
           </p>
           <input
             style={{
@@ -154,9 +163,10 @@ const Register = () => {
               fontSize: 16,
               borderRadius: 10,
               marginBottom: "1em",
+              border:"1px solid grey"
             }}
             onChange={(e) => {
-              setUser({ ...user, password: e.target.value });
+              setUser({ ...user, confirmPassword: e.target.value });
             }}
           />
           <p
